telegram_ucp_project/gaura_platform/requirements.txt
fastapi
uvicorn
python-telegram-bot
sqlalchemy
httpx
pydantic
pydantic-settings
google-generativeai
python-dotenv
supabase
pytz
telethon
razorpay
openai
